requirements.txt
streamlit>=1.28.0
youtube-transcript-api>=0.6.1
PyMuPDF>=1.23.0
python-pptx>=0.6.21
langchain>=0.1.0
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
